refactor(smt): rename rocksdb.rs to persistent.rs
After generic KVDBSmtStorage refactoring this file is not tied to RocksDB.
