+++ miden-crypto/src/merkle/smt/large/storage/mod.rs
@@ -25,14 +25,14 @@ pub mod config;
 mod schema;
 
 #[cfg(feature = "smt-kvdb")]
-mod rocksdb;
+mod persistent;
 
 #[cfg(feature = "smt-kvdb-rocks")]
 mod rocks_kvdb;
 #[cfg(feature = "smt-kvdb-rocks")]
-pub type PersistentSmtStorage = rocksdb::KVDBSmtStorage<rocks_kvdb::RocksKVDB>;
+pub type PersistentSmtStorage = persistent::KVDBSmtStorage<rocks_kvdb::RocksKVDB>;
 #[cfg(feature = "smt-kvdb-rocks")]
-pub type PersistentSmtStorageSnapshot = rocksdb::KVDBSnapshotStorage<rocks_kvdb::RocksKVDB>;
+pub type PersistentSmtStorageSnapshot = persistent::KVDBSnapshotStorage<rocks_kvdb::RocksKVDB>;
 
 mod memory;
 pub use memory::{MemoryStorage, MemoryStorageSnapshot};
